requirements.txt
asgiref==3.5.2
beautifulsoup4==4.9.3
boto==2.46.1
bs4==0.0.1
colorlog==3.1.4
chardet==3.0.4
crossrefapi==1.5.0
Django==3.2.20
django-bleach==3.0.1
django-bootstrap4==22.3
-e git+https://github.com/BirkbeckCTP/django-foundation-form.git@284fc5f08e58d8655c7301eddb49f920fda516d2#egg=foundationform
django-countries[pyuca]==7.5.1
-e git+https://github.com/kalwalkden/django-materializecss-form.git@7018fe0659ad12e74cfea80e281faa1390a3132f#egg=django-materializecss-form
django-debug-toolbar==3.7.0
django-hCaptcha==0.2.0
django-hijack==3.2.1
-e git+https://github.com/BirkbeckCTP/django-mailgun.git@8f97af5a4133934f939826c00a30f1357478b139#egg=django_mailgun
django-modeltranslation==0.18.9
django-settings-export==1.2.1
django-recaptcha==3.0.0
django-simple-math-captcha==1.1.0
django-simple-history==3.3.0
django-summernote==0.8.20.0
djangorestframework==3.11.2
-e git+https://github.com/BirkbeckCTP/django-dynamicsites@f439c524cf964d957a75fbec4b23ad03a9a5d159#egg=dynamicsites
-e git+https://github.com/BirkbeckCTP/ebooklib@b7005c57602eeea06358ac569a4cf52c630a32ef#egg=ebooklib
Faker==17.6.0
freezegun==1.2.0
geoip2==3.0.0 # Newer versions require python >= 3.6
html2text==2017.10.4
idna==2.7
ipaddress==1.0.16
more-itertools==10.1.0
jsmin==3.0.1
kitchen==1.2.4
libsass==0.22.0
lxml==4.9.2
markdown
maxminddb==1.5.4
mock
mozilla-django-oidc==1.2.4
orcid==1.0.3
packaging==23.2
pdfkit==1.0.0
pdfminer.six
psycopg2-binary~=2.8.6
pyasn1==0.1.9
pycparser==2.14
PyJWT==2.4.0
Pillow==10.0.1
python-crontab==2.2.3
python-dateutil==2.8.1
python-docx==0.8.11
python-magic==0.4.13
requests==2.31.0
six==1.14.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.65.0
ua-parser==0.16.1
# See #3736
urllib3<2
user-agents==2.2.0
